Relocate NewBoardLink component
