+++ src/app/[locale]/dashboard/_components/Topbar/Topbar.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import styles from './styles';
-import NewBoardLink from '../../[id]/NewBoardLink/NewBoardLink';
+import NewBoardLink from '../NewBoardLink/NewBoardLink';
 import OpenSideBarButton from '../OpenSideBarButton/OpenSideBarButton';
 
 type TopbarProps = {
